divide ClientProperties into TmProperties and RmProperties,
