+++ script/client/spring/application.properties
@@ -19,18 +19,18 @@ seata.application-id=test
 seata.tx-service-group=my_test_tx_group
 seata.enable-auto-data-source-proxy=true
 seata.use-jdk-proxy=false
-seata.client.rm-async-commit-buffer-limit=1000
-seata.client.rm-report-retry-count=5
-seata.client.rm-table-meta-check-enable=false
-seata.client.rm-report-success-enable=true
-seata.client.rm.lock.lock-retry-interval=10
-seata.client.rm.lock.lock-retry-times=30
-seata.client.rm.lock.lock-retry-policy-branch-rollback-on-conflict=true
-seata.client.tm-commit-retry-count=5
-seata.client.tm-rollback-retry-count=5
-seata.client.undo.undo-data-validation=true
-seata.client.undo.undo-log-serialization=jackson
-seata.client.undo.undo-log-table=undo_log
+seata.client.rm.async-commit-buffer-limit=1000
+seata.client.rm.report-retry-count=5
+seata.client.rm.table-meta-check-enable=false
+seata.client.rm.report-success-enable=true
+seata.client.rm.lock.retry-interval=10
+seata.client.rm.lock.retry-times=30
+seata.client.rm.lock.retry-policy-branch-rollback-on-conflict=true
+seata.client.tm.commit-retry-count=5
+seata.client.tm.rollback-retry-count=5
+seata.client.undo.data-validation=true
+seata.client.undo.log-serialization=jackson
+seata.client.undo.log-table=undo_log
 seata.client.log.exceptionRate=100
 seata.service.vgroup-mapping=default
 seata.service.grouplist=127.0.0.1:8091
